Ideally a logging framework doesn't cause a panic for code like this:

```go
foo, err := getFoo()

if err != nil {
   log.Error("Couldn't acquire foo", zap.Error(err))
}

// foo could be nil here
log = log.With(zap.Stringer("foo", foo))

bar, err := getBar()

if err != nil {
  // panics
  log.Error("Couldn't acquire bar", zap.Error(err))
}
```

This is probably also fixing #495 and #528