@prashantv 2019-03-05
Thanks for the contribution @joa.

Your solution works when dealing with functions that return interfaces, and explicitly return `nil`. However, it doesn't work when the return value has a type, but the value is `nil`, as mentioned in https://github.com/uber-go/zap/issues/495#issuecomment-324805952

With your PR merged in, I wrote a simple test:
```go
func getURL() (*url.URL, error) {
  return nil, nil
}

url, err := getURL()
if err != nil {
  logger.Error("got error", zap.Error(err))
}

logger.Info("got url", zap.Stringer("url", url))
```

This still causes a panic:
```
panic: runtime error: invalid memory address or nil pointer dereference [recovered]
        panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x8 pc=0x5d3550]

(truncated)

panic(0x74d580, 0x9bc7d0)
        /home/prashant/bin/go1.10.4/src/runtime/panic.go:502 +0x229
net/url.(*URL).String(0x0, 0x790a00, 0x0)
        /home/prashant/bin/go1.10.4/src/net/url/url.go:742 +0x40
go.uber.org/zap/zapcore.Field.AddTo(0x7b73a5, 0x3, 0x18, 0x0, 0x0, 0x0, 0x790a00, 0x0, 0x81ed20, 0xc42009d170)
        /home/prashant/go/src/go.uber.org/zap/zapcore/field.go:172 +0xc10
```

See related questions:
https://golang.org/doc/faq#nil_error
https://stackoverflow.com/questions/38070015/check-if-interface-value-is-nil-in-go-without-using-reflect/38070426