@joa 2019-03-05
Good catch. I expanded the test case and added the additional nil check using reflection. 

I was unsuccessful with other variants. I hacked a bit around `unsafe.Pointer` but that's also not promising. Each `(type, value)` tuple is a new instance and you can't compare the heap address for anything.