@prashantv 2019-03-06
Unfortunately reflection comes with a performance penalty, and the only alternative is `unsafe`, which we strongly discourage.

It's also not correct to assume that `nil` always means that the type cannot be logged. E.g., it's completely reasonable to have a type which has a `String()` method that handles the `nil` case.