@joa 2019-03-06
@prashantv The commit does not use reflection to test for the nil pointer in side the non-nil `(type, ptr)` tuple. Do you have a particular reason why you'd advise against unsafe in this instance? There are unit tests in place that cover both the positive and negative cases.  Of course the compatibility promise isn't given when using unsafe but that's why the tests exist.

I understand the argument around a valid `String()` implementation for the nil case. In fact it is best practice to provide meaningful nil values and have these cases covered. However as with `url.URL` we see this isn't even consistent in within the go lib itself. 

Of course both cases could be supported with `recover` and accepting a potential panic. I guess however I'd rather have a different layer catch these (e.g. a middleware in a server). I guess there is nothing left in this MR to improve directly. Feel free to close if you think the risks outweigh the benefits.