@joa 2019-03-06
I see from your comment in https://github.com/uber-go/zap/issues/495#issuecomment-326355875 that fmt is acutally doing this. Maybe that'd be an acceptable solution.