@joa 2019-03-06
With defer/recover
```
BenchmarkAddingFields/Zap-8                                  	 1000000	      2212 ns/op	    2214 B/op	      15 allocs/op
BenchmarkAddingFields/Zap.Check-8                            	 1000000	      2221 ns/op	    2214 B/op	      15 allocs/op
BenchmarkAddingFields/Zap.CheckSampled-8                     	 5000000	       312 ns/op	     277 B/op	       1 allocs/op
BenchmarkAddingFields/Zap.Sugar-8                            	  500000	      2762 ns/op	    3010 B/op	      21 allocs/op
BenchmarkAddingFields/apex/log-8                             	   50000	     33792 ns/op	    6694 B/op	     115 allocs/op
BenchmarkAddingFields/go-kit/kit/log-8                       	  200000	      8049 ns/op	    5641 B/op	     105 allocs/op
BenchmarkAddingFields/inconshreveable/log15-8                	   50000	     30263 ns/op	    6625 B/op	     122 allocs/op
BenchmarkAddingFields/sirupsen/logrus-8                      	  200000	      9931 ns/op	    9082 B/op	     122 allocs/op
BenchmarkAddingFields/go.pedge.io/lion-8                     	  200000	      8117 ns/op	    8761 B/op	     105 allocs/op
BenchmarkAddingFields/rs/zerolog-8                           	  200000	      6691 ns/op	    4725 B/op	      76 allocs/op
BenchmarkAddingFields/rs/zerolog.Check-8                     	  200000	      6646 ns/op	    4728 B/op	      76 allocs/op
```

Without defer/recover:
```
BenchmarkAddingFields/Zap-8                                  	 1000000	      1614 ns/op	     770 B/op	       5 allocs/op
BenchmarkAddingFields/Zap.Check-8                            	 1000000	      1614 ns/op	     770 B/op	       5 allocs/op
BenchmarkAddingFields/Zap.CheckSampled-8                     	10000000	       225 ns/op	      91 B/op	       0 allocs/op
BenchmarkAddingFields/Zap.Sugar-8                            	 1000000	      1986 ns/op	    1565 B/op	      11 allocs/op
BenchmarkAddingFields/apex/log-8                             	   50000	     33770 ns/op	    6694 B/op	     115 allocs/op
BenchmarkAddingFields/go-kit/kit/log-8                       	  200000	      8100 ns/op	    5640 B/op	     105 allocs/op
BenchmarkAddingFields/inconshreveable/log15-8                	   50000	     30802 ns/op	    6623 B/op	     122 allocs/op
BenchmarkAddingFields/sirupsen/logrus-8                      	  200000	     10017 ns/op	    9082 B/op	     122 allocs/op
BenchmarkAddingFields/go.pedge.io/lion-8                     	  200000	      8220 ns/op	    8760 B/op	     105 allocs/op
BenchmarkAddingFields/rs/zerolog-8                           	  200000	      6620 ns/op	    4728 B/op	      76 allocs/op
BenchmarkAddingFields/rs/zerolog.Check-8                     	  200000	      6662 ns/op	    4725 B/op	      76 allocs/op
```