@joa 2019-03-06
Implemented recover within defer and expanded test cases as discussed (e.g. valid nil impl for Stringer).