@prashantv 2019-03-25
Thanks @joa. To better understand the performance impact, I wrote a benchmark which:
 * Uses a production logger config, writing to `/dev/null`
 * Created a custom stringer type which doesn't need to encode anything

E.g.,
```go
type customString string

func (cs customString) String() string {
        return string(cs)
}

func BenchmarkStringer(b *testing.B) {
        cfg := zap.NewProductionConfig()
        cfg.OutputPaths = []string{"/dev/null"}
        l, err := cfg.Build()
        require.NoError(b, err, "failed to build logger")

        cs := customString("encode")

        b.Run("1 field", func(b *testing.B) {
                for i := 0; i < b.N; i++ {
                        l.Info("test",
                                zap.Stringer("cs1", cs),
                        )
                }
        })
        b.Run("4 fields", func(b *testing.B) {
                for i := 0; i < b.N; i++ {
                        l.Info("test",
                                zap.Stringer("cs1", cs),
                                zap.Stringer("cs2", cs),
                                zap.Stringer("cs3", cs),
                                zap.Stringer("cs4", cs),
                        )
                }
        })
}
```

To get a closer look at the impact to just the stringer performance, I wrote a benchmark which only logs stringer fields,
```
        b.Run("1 field", func(b *testing.B) {
                for i := 0; i < b.N; i++ {
                        l.Info("test",
                                zap.Stringer("cs1", cs),
                        )
                }
        })
        b.Run("4 fields", func(b *testing.B) {
                for i := 0; i < b.N; i++ {
                        l.Info("test",
                                zap.Stringer("cs1", cs),
                                zap.Stringer("cs2", cs),
                                zap.Stringer("cs3", cs),
                                zap.Stringer("cs4", cs),
                        )
                }
        })
```

I then compared the difference with and without this change on Go 1.12,

Without this change (no `defer`):
```
BenchmarkStringer/1_field-4             20000000               203 ns/op              82 B/op          2 allocs/op
BenchmarkStringer/4_fields-4            20000000               373 ns/op             322 B/op          5 allocs/op
```

With this change (with `defer`):
```
BenchmarkStringer/1_field-4             20000000               207 ns/op              83 B/op          2 allocs/op
BenchmarkStringer/4_fields-4            10000000               377 ns/op             327 B/op          5 allocs/op
```

There's a very slight increase in the number of bytes allocated, but the actual number of allocations remains the same, and the performance is the same (the difference looks like noise).